# backend/services/crm/src/modules/contacts/contacts.controller.ts
import { Controller, Get, Post, Patch, Delete, Body, Param, Query, UseGuards } from '@nestjs/common';
import { JwtAuthGuard } from '@eventos-ai/auth';
import { ContactsService } from './contacts.service';

@Controller('crm/contacts')
@UseGuards(JwtAuthGuard)
export class ContactsController {
    constructor(private readonly service: ContactsService) {}

    @Post()
    async create(@Body() body: any) { return this.service.create(body); }

    @Get()
    async findAll(
        @Query('organizationId') orgId: string,
        @Query('search') search?: string,
        @Query('page') page?: number,
        @Query('perPage') perPage?: number,
    ) { return this.service.findAll(orgId, { search, page, perPage }); }

    @Get(':id')
    async findById(@Param('id') id: string) { return this.service.findById(id); }

    @Patch(':id')
    async update(@Param('id') id: string, @Body() body: any) { return this.service.update(id, body); }

    @Delete(':id')
    async remove(@Param('id') id: string) { return this.service.remove(id); }
}